requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
CacheControl==0.14.3
cachetools==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.4
daphne==4.2.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
exceptiongroup==1.3.0
firebase-admin==6.9.0
flower==2.0.1
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.4.4
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==2.0.7
python-telegram-bot==13.15
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.3.0
tornado==6.1
twilio==9.6.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
zope.interface==7.2
drf-spectacular==0.27.1
